backend/pyproject.toml
[tool.poetry]
name = "restaurant-finder-backend"
version = "0.1.0"
description = "Backend for the Restaurant Finder application using Google Maps MCP and LLaMA"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.109.2"
uvicorn = "0.27.1"
pydantic = "2.6.1"
python-dotenv = "1.0.1"
requests = "2.31.0"
openai = "1.12.0"
google_maps_agent = { path = "../google_maps_agent" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
